requirements.txt
streamlit
edgartools
pandas
plotly
fpdf2